# api/rest/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Chapar Mock Server REST API
  description: >-
    Mock API for exercising HTTP clients such as Chapar. It exposes an
    in-memory todo resource and a set of utility endpoints (echo, status codes,
    delays, auth schemes, uploads, redirects, cookies, streaming).


    Todos are scoped to a session. Send the same `X-Session-Id` header on every
    request to get an isolated list; requests without it share the `public`
    session. Sessions start with a few sample todos and expire after a period
    of inactivity.
servers:
  - url: https://mocks.chapar.rest/api/v1
  - url: http://localhost:8080/api/v1
...
paths:
...
  /bytes/{n}:
    get:
      operationId: GetBytes
      tags: [utility]
      summary: Return n random bytes
      parameters:
        - name: n
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 102400
      responses:
        "200":
          description: Random bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"